requirements.txt
asgiref==3.5.2
bokeh==2.4.2
charset-normalizer==2.0.12
Django==4.0
django-crontab==0.7.1
django-environ==0.8.1
django-extensions==3.1.3
django-filter==22.1
django-htmx==1.6.0
django-widget-tweaks==1.4.9
djangorestframework==3.14.0
idna==3.4
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
pytz==2022.6
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.3
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
wincertstore==0.2
